+++ libcaliptra/inc/caliptra_api.h
@@ -5,10 +5,12 @@
 #include <stdbool.h>
 
 #include "caliptra_types.h"
-#include "caliptra_enums.h"
-#include "caliptra_if.h"
 
 #define CALIPTRA_ARRAY_SIZE(array) ((size_t)(sizeof(array) / sizeof(array[0])))
+
+#ifdef __cplusplus
+extern "C" {
+#endif
 
 // Write into Caliptra BootFSM Go Register
 // returns: 0                       - Success
@@ -36,7 +38,7 @@ bool caliptra_ready_for_fuses(void);
 //          INVALID_PARAMS          - Pointer to fuse struct is NULL
 //          NOT_READY_FOR_FUSES     - Flow status does not indicate ready for fuses before initializing
 //          STILL_READY_FOR_FUSES   - Flow status still indicates ready for fuses after writing fuse done
-int caliptra_init_fuses(struct caliptra_fuses *fuses);
+int caliptra_init_fuses(const struct caliptra_fuses *fuses);
 
 // Query if ROM is ready for firmware
 bool caliptra_ready_for_firmware(void);
@@ -92,12 +94,12 @@ int caliptra_complete();
 // Generic function for sending and receiving a command with raw, user-defined command and buffers
 // NOT RECOMMENDED to be used directly when avoidable - use the functions below for the specific command
 // (See notes above on asynchronous operation and return codes)
-int caliptra_mailbox_execute(uint32_t cmd, struct caliptra_buffer *mbox_tx_buffer, struct caliptra_buffer *mbox_rx_buffer, bool async);
+int caliptra_mailbox_execute(uint32_t cmd, const struct caliptra_buffer *mbox_tx_buffer, struct caliptra_buffer *mbox_rx_buffer, bool async);
 
 // For full command details, please refer to the Caliptra Runtime Readme file at runtime\README.md
 
 // Upload Caliptra Firmware
-int caliptra_upload_fw(struct caliptra_buffer *fw_buffer, bool async);
+int caliptra_upload_fw(const struct caliptra_buffer *fw_buffer, bool async);
 
 // Get IDEV cert
 int caliptra_get_idev_cert(struct caliptra_get_idev_cert_req *req, struct caliptra_get_idev_cert_resp *resp, bool async);
@@ -180,4 +182,8 @@ void caliptra_req_idev_csr_start();
 
 // Clear IDEV CSR request. 
 void caliptra_req_idev_csr_complete(); 
+
+#ifdef __cplusplus
+}
+#endif
