+++ libcaliptra/src/caliptra_fuses.h
@@ -2,13 +2,21 @@
 
 #pragma once
 
-#include "caliptra_api.h"
+#include <stddef.h>
+#include <stdint.h>
+
+#include <caliptra_top_reg.h>
+#include "caliptra_if.h"
 
 #define MBOX_PAUSER_SLOTS (5)
 
 // WARNING: THESE APIS ARE INTENDED FOR SIMULATION ONLY.
 //          SOC FW MUST HAVE NO ACCESS TO THOSE APIS.
 //          A HW STATE MACHINE SHOULD BE USED TO SEND FUSE VALUES TO CALIPTRA OVER APB BUS
+
+#ifdef __cplusplus
+extern "C" {
+#endif
 
 static inline void caliptra_generic_and_fuse_write(uint32_t offset, uint32_t data)
 {
@@ -22,7 +30,7 @@ static inline uint32_t caliptra_generic_and_fuse_read(uint32_t offset)
     return data;
 }
 
-static inline void caliptra_fuse_array_write(uint32_t offset, uint32_t *data, size_t size)
+static inline void caliptra_fuse_array_write(uint32_t offset, const uint32_t *data, size_t size)
 {
     for (uint32_t idx = 0; idx < size; idx ++)
     {
@@ -116,3 +124,8 @@ static inline void caliptra_write_dbg_manuf_serv(uint32_t data)
     // Set Manuf service reg
     caliptra_generic_and_fuse_write(GENERIC_AND_FUSE_REG_CPTRA_DBG_MANUF_SERVICE_REG, data);    
 }
+
+
+#ifdef __cplusplus
+}
+#endif
